Add Cubase Input Transformer rules for Arturia V Collection use of relative mode.
Arturia V Collection implementation of controller relative mode uses a nonstandard MIDI encoding. Arturia has acknowledged this but gave no input of if or when it will be corrected.

In the meantime, you will need to add these two Input Transformer entries to each Cubase track using a V Collection soft instrument. Another option is to modify the JavaScript to place the controls into absolute mode. The default is relative mode as this is considered the best solution for continuous rotary encoders to avoid jumps in value.